+++ src/main/java/me/sv3r/cutehermitcrabs/common/item/HermitCrabShellItem.java
@@ -1,5 +1,6 @@
 package me.sv3r.cutehermitcrabs.common.item;
 
+import me.sv3r.cutehermitcrabs.common.config.CHCConfig;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TranslatableComponent;
@@ -30,7 +31,7 @@ public class HermitCrabShellItem extends ArmorItem
     {
         if (player.isShiftKeyDown())
         {
-            player.addEffect(new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, 10, 2, false, false, true));
+            player.addEffect(new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, 10, CHCConfig.RESISTANCE_LEVEL.get() - 1, false, false, true));
         }
     }
 
@@ -40,7 +41,7 @@ public class HermitCrabShellItem extends ArmorItem
         tooltipComponents.add(new TranslatableComponent("item.modifiers.sneaking").withStyle(ChatFormatting.GRAY));
         tooltipComponents.add(new TranslatableComponent("attribute.modifier.plus").withStyle(ChatFormatting.BLUE)
                 .append(new TranslatableComponent(MobEffects.DAMAGE_RESISTANCE.getDisplayName().getString()).withStyle(ChatFormatting.BLUE)).append(" ")
-                .append(new TranslatableComponent("potion.potency.2")).withStyle(ChatFormatting.BLUE));
+                .append(new TranslatableComponent("potion.potency." + (CHCConfig.RESISTANCE_LEVEL.get() - 1))).withStyle(ChatFormatting.BLUE));
 
         super.appendHoverText(itemStack, level, tooltipComponents, isAdvanced);
     }