+++ src/main/java/me/sv3r/cutehermitcrabs/common/misc/CHCFoodValues.java
@@ -1,8 +1,8 @@
-package me.sv3r.cutehermitcrabs.common.registry;
+package me.sv3r.cutehermitcrabs.common.misc;
 
 import net.minecraft.world.food.FoodProperties;
 
-public class ModFoodValues
+public class CHCFoodValues
 {
     // Meats
     public static final FoodProperties RAW_HERMIT_CRAB = (new FoodProperties.Builder()).nutrition(2).saturationMod(0.4f).build();