+++ CHANGELOG.md
@@ -2,14 +2,23 @@
 
 ## 2.0.2
 
+### Additions
+
+- Added a config file for changing hermit crab spawn values & hermit crab shell resistance level;
+
 ### Fixes
-- Fixed smelter not being able to cook raw hermit crab;
+
+- Fixed smoker not being able to cook raw hermit crab;
+- Fixed entity.cutehermitcrabs.hermit_crab not being translated;
+- Fixed bucket of hermit crab not saving name & health;
 
 ## 2.0.1
 
 ### Additions
+
 - Added nl_NL translation;
 - Added nl_BE translation;
 
 ### Fixes
+
 - Fixed mod crashing servers;